ステルスキルの判定を追加

+++ Project/SourceCode/Part/graphicer.hpp
@@ -26,7 +26,7 @@ public:
 	[[nodiscard]] Vector2D<float>	GetScale()				const { return m_scale; }
 	[[nodiscard]] float				GetAngle()				const { return m_angle; }
 	[[nodiscard]] Vector2D<int>		GetOriginSize()			const { return Vector2D<int>(m_origin_width, m_origin_height); }
-	[[nodiscard]] Vector2D<int>		GetHalfOriginSize()		const { return Vector2D<int>(m_origin_width * 0.5f, m_origin_height * 0.5f); }
+	[[nodiscard]] Vector2D<int>		GetHalfOriginSize()		const { return Vector2D<int>(static_cast<int>(m_origin_width * 0.5f), static_cast<int>(m_origin_height * 0.5f)); }
 	[[nodiscard]] int				GetBlendMode()			const { return m_blend_mode; }
 	[[nodiscard]] int				GetBlendNum()			const { return m_blend_num; }
 	[[nodiscard]] bool				IsInvertHorizontal()	const { return m_is_invert_horizontal; }
